requirements.txt
azure_storage==0.37.0
python-dotenv==0.21.0
